pyproject.toml
[project]
name = "the-evolution-of-todo"
version = "0.1.0"
description = "The Evolution of Todo - From CLI to Full-Stack"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.127.0",
    "uvicorn>=0.40.0",
    "sqlmodel>=0.0.29",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sse-starlette>=2.1.3",
    "websockets>=14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
]

[tool.uv]
package = false

[tool.uv.workspace]
members = [
    "src/backend",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]
